fix(phase-2): fix untyped-NULL parameter and cross-file test collision
CI caught two real bugs:

1. resolve_location's INSERT used CASE WHEN %(longitude)s IS NULL ...
   ELSE ST_SetSRID(ST_MakePoint(%(longitude)s, %(latitude)s), 4326)
   END. When longitude/latitude are Python None (the ambiguous,
   no_match, out_of_bounds, and invalid_input branches), psycopg sends
   an untyped NULL and Postgres cannot infer a type for the parameter
   from an IS NULL check alone, raising AmbiguousParameter. The
   accepted-status test passed locally because a real Python float
   gives psycopg a concrete type to adapt from — the bug only shows up
   on the NULL path, which none of the local (non-integration) tests
   exercise. Fixed with explicit ::float8 casts on both parameters.

2. test_resolve_location_accepted_with_full_region_hierarchy asserted
   result.regions had *exactly* the keys {sa2, lga, poa}, but
   resolve_point checks every region_type against whatever is active
   for it across the whole shared CI database - test_asgs.py's own
   sa3/sa4 tests happen to use the identical Sydney-area coordinates
   (151.0-151.1, -33.8/-33.9) this file's fixtures used, so their
   still-active regions genuinely (and correctly) also matched this
   test's point. Not a resolve_point bug - two different test files'
   fixtures coincidentally overlapped in the one dimension that
   matters here: real-world geographic coordinates are shared, global
   state, unlike a random UUID suffix. Fixed by moving this file's
   fixtures to a distinct Perth-area footprint instead of loosening the
   assertion, so the next test file added won't silently reintroduce
   the same collision by picking yet another coincidentally-close
   Sydney coordinate.

+++ workers/ingestion/src/austechmap_ingestion/geography/locations.py
@@ -189,8 +189,9 @@ def resolve_location(
             )
             VALUES (
               %(input_hash)s, %(input_text)s, %(status)s, %(method)s,
-              CASE WHEN %(longitude)s IS NULL THEN NULL
-                   ELSE ST_SetSRID(ST_MakePoint(%(longitude)s, %(latitude)s), 4326) END,
+              CASE WHEN %(longitude)s::float8 IS NULL THEN NULL
+                   ELSE ST_SetSRID(ST_MakePoint(%(longitude)s::float8, %(latitude)s::float8), 4326)
+                   END,
               %(sa1)s, %(sa2)s, %(sa3)s, %(sa4)s, %(lga)s, %(poa)s,
               %(migration_category)s, %(migration_dama_name)s
             )
